@@ -7,31 +7,34 @@
     "start:server": "ts-node-dev scripts/test-server.ts"
   },
   "dependencies": {
-    "@babel/core": "^7.7.5",
-    "@babel/preset-env": "^7.7.6",
-    "@babel/preset-react": "^7.7.4",
-    "@data-proxy/react-admin-provider": "^0.10.2",
-    "@data-proxy/server": "^0.11.3",
-    "@material-ui/core": "^4",
-    "@material-ui/icons": "^4",
-    "@react-admin-git-provider/gitlab": "^0.9.0",
-    "babel-loader": "^8.0.6",
+    "@babel/core": "^7.24.3",
+    "@babel/preset-env": "^7.24.3",
+    "@babel/preset-react": "^7.24.1",
+    "@mui/icons-material": "^5.15.14",
+    "@mui/material": "^5.15.14",
+    "@react-admin-git-provider/gitlab": "^0.13.0",
+    "babel-loader": "^9.1.3",
     "babel-polyfill": "^6.26.0",
-    "html-webpack-plugin": "^3.2.0",
-    "react": "^16.12.0",
-    "react-admin": "^3.0.4",
-    "react-dom": "^16.12.0",
-    "webpack": "^4.41.2",
-    "webpack-cli": "^3.3.10",
-    "webpack-dev-server": "^3.9.0"
+    "fs-extra": "^11.2.0",
+    "html-webpack-plugin": "^5.6.0",
+    "react": "^18.2.0",
+    "react-admin": "^4.16.13",
+    "react-dom": "^18.2.0",
+    "webpack": "^5.91.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^5.0.4"
   },
   "devDependencies": {
     "@data-proxy/server-auth-omnipartners": "^0.9.10",
+    "@types/fs-extra": "^11.0.4",
+    "@types/react-dom": "^18.2.22",
     "babel-plugin-transform-class-properties": "^6.24.1",
-    "express": "^4.17.1",
-    "lodash": "^4.17.15",
-    "morgan": "^1.9.1",
-    "omnipartners": "^2.21.0",
-    "ts-node-dev": "^1.0.0-pre.43"
+    "express": "^4.19.2",
+    "lodash": "^4.17.21",
+    "morgan": "^1.10.0",
+    "omnipartners": "^2.92.27",
+    "ts-loader": "^9.5.1",
+    "ts-node-dev": "^2.0.0",
+    "typescript": "^5.4.3"
   }
 }
