@@ -7,29 +7,36 @@ import {
 import { LocalforageCacheProvider } from "@react-admin-git-provider/common";
 
 const baseProviderOptions = {
-  projectId: process.env.GITLAB_PROJECT_ID,
-  ref: process.env.GITLAB_REF,
+  projectId: process.env.GITLAB_PROJECT_ID || "",
+  ref: process.env.GITLAB_REF || "",
   gitlabOptions: {
-    host: process.env.REACT_ADMIN_DATA_API,
+    host: process.env.REACT_ADMIN_DATA_API || "",
   },
 };
 
-export const getProviderForResource = resource =>
+export const getProviderForResource = (resource: string) =>
   resource === "pipelines"
     ? new GitlabProviderPipeline({
         ...baseProviderOptions,
+        path: `${process.env.GITLAB_DATA_BASE_PATH}/${resource}.json`,
       })
     : resource === "articles"
     ? new GitlabProviderFile({
         ...baseProviderOptions,
         path: `${process.env.GITLAB_DATA_BASE_PATH}/${resource}.json`,
         cacheBehavior: "contentSha",
-        cacheProvider: new LocalforageCacheProvider({ storeName: "react-admin-gitlab" }),
+        cacheProvider: new LocalforageCacheProvider({
+          storeName: "react-admin-gitlab",
+        }),
+        serializer: "json",
       })
     : new GitlabProviderFileList({
         ...baseProviderOptions,
         path: `${process.env.GITLAB_DATA_BASE_PATH}/${resource}`,
-        cacheProvider: new LocalforageCacheProvider({ storeName: "react-admin-gitlab" }),
+        cacheProvider: new LocalforageCacheProvider({
+          storeName: "react-admin-gitlab",
+        }),
+        serializer: "json",
       });
 
 export const dataProvider = createDataProvider(({ resource }) =>
